fix the Base api url on production error (but i secretly destroy your Dev Testing :D good luck))
when on the production the base api calling "localhost" this  is the issue.
what change is from calling with "localhost" now on production calling with empty string "" this forge the client to use the DNS load banalcer name to communicate with server. but the side effect is I just mess your dev testing environment now it doesnt work on environment anymore hahahahahaha